+++ test_unstructured_api_tools/pipeline-test-project/prepline_test_project/api/process_file_4.py
@@ -82,7 +82,11 @@ def get_validated_mimetype(file):
 
         if content_type not in allowed_mimetypes:
             raise HTTPException(
-                status_code=400, detail=f"File type not supported: {file.filename}"
+                status_code=400,
+                detail=(
+                    f"Unable to process {file.filename}: "
+                    f"File type {content_type} is not supported."
+                ),
             )
 
     return content_type