+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+# 0.9.2
+
+* Add content_type to error message for unsupported file types
+  
+# 0.9.1
+
+* Allow references to standard imports in pipeline cells
+
 # 0.9.0
 
 * Add supporting gzip compressed files
-
 
 # 0.8.1